+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # UNRELEASED 
 
 * Can now click between cells to create a new cell [#235]
+* New clear output button - deletes all results/outputs of a notebook from UI and underlying file [#237]
 
 [#235]: https://github.com/polynote/polynote/issues/235
+[#237]: https://github.com/polynote/polynote/issues/237
 
 # 0.1.4 (April 16, 2019)